arm: Ambarella: GPIO: Fix locking in gpio_get/set/configure
The ambarella_gpio_inline_config() function will already
take the GPIO chip lock prior to writing to hardware
registers, so attempting to also lock the GPIO chip at the
call site to ambarella_gpio_inline_config() will inevitably
lead to a deadlock.

Moreover, the locking code tried to index the per-bank
locks by GPIO number, which is not valid and will lead to
all manner of badness. So, drop the redundant locking.

Signed-off-by: evilwombat <evilwombat@server.fake>