mtd: nand: Ambarella: Assign a name to our MTD
Assign a name to the 'name' field to the ambarella_nand MTD
structure, or the NAND layer will assign the name to match
the NAND part description string, which is not what we want.
A generic name (containing no spaces, at that) allows the
MTD to be targeted for commandline partitioning, which is
handy if we want to bypass a damaged PTB while keeping the
firmware region separate from the Cal region.

Signed-off-by: evilwombat <evilwombat@server.fake>