leds: as3715: Don't write backlight registers without ST7789
To enable backlight control on the Hero4 Silver, the AS3715
PMIC LED driver currently initializes the PMIC registers to
a "known good" state. However, some of these registers
perform other useful functions, such as enabling access to
the SD card slot, while others may not be appropriate on a
Hero4 Black camera, due to lack of a backlight.

To allow camera recovery to work properly on Hero4 Black
and Silver, we still need to write to a subset of the PMIC
registers to enable the SD slot. However, we need to avoid
writing to unrelated registers in this case (and if support
for ST7789 is compiled out, we don't care about the
backlight anyway).

Through trial and error, we find the register writes that
are seemingly needed for SD card access, and only perform
these writes if we are running in recovery mode (without
display).

Signed-off-by: evilwombat <evilwombat@server.fake>